Join test infra CLI and GUI
