+++ tests/config.js
@@ -1,0 +1,31 @@
+const IDFVERSIONSSHOWNONINSTALLER = [
+    `master`,
+    `v5.4.1`,
+    `v5.3.2`,
+    `v5.2.3`,
+    `v5.1.6`,
+];
+
+const IDFMIRRORS = {
+    github: "https://github.com",
+    jihulab: "https://jihulab.com/esp-mirror",
+};
+const TOOLSMIRRORS = {
+    github: "https://github.com",
+    dl_com: "https://dl.espressif.com/github_assets",
+    dl_cn: "https://dl.espressif.cn/github_assets",
+};
+
+const CLIDEFAULTVERSION = "eim 0.2.0";
+const GUIDEFAULTVERSION = "0.1.2";
+
+const IDFDEFAULTINSTALLVERSION = "v5.4.1";
+
+export {
+    IDFVERSIONSSHOWNONINSTALLER,
+    IDFMIRRORS,
+    TOOLSMIRRORS,
+    CLIDEFAULTVERSION,
+    GUIDEFAULTVERSION,
+    IDFDEFAULTINSTALLVERSION,
+};
