+++ tests/suites/CLI-prerequisites.json
@@ -1,0 +1,7 @@
+[
+    {
+        "id": 1,
+        "type": "prerequistes",
+        "name": "CLI-prerequisites"
+    }
+]
