--- tests_cli/package.json
@@ -1,17 +1,0 @@
-{
-    "type": "module",
-    "dependencies": {
-        "chai": "^5.1.1",
-        "mocha": "^10.7.3",
-        "mocha-junit-reporter": "^2.2.1",
-        "node-gyp": "^10.2.0",
-        "node-pty": "^1.0.0",
-        "winston": "^3.15.0"
-    },
-    "scripts": {
-        "pre-test": "mocha --exit --reporter json --reporter-options output=./results-pre-requisites-test.json script/prerequisites.test.js",
-        "pre-install": "mocha --exit --bail --reporter json --reporter-options output=./results-pre-install-test.json script/prerequisitesInstallRunner.test.js",
-        "test": "export JSON_FILENAME=$npm_config_file && mocha --exit --reporter json --reporter-options output=./results-$JSON_FILENAME.json runs/testRun.test.js",
-        "test-win": "powershell -Command \"$env:JSON_FILENAME=$env:npm_config_file; mocha --exit --reporter json --reporter-options output=./results-$env:JSON_FILENAME.json runs/testRun.test.js\""
-    }
-}
